+++ utils/utils.c
@@ -6,13 +6,13 @@
 /*   By: msokolov <msokolov@student.codam.nl>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/05 23:57:55 by msokolov          #+#    #+#             */
-/*   Updated: 2025/12/06 13:31:10 by msokolov         ###   ########.fr       */
+/*   Updated: 2025/12/11 22:48:23 by msokolov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-int	get_map_height(char **map)
+size_t get_map_height(char **map)
 {
 	int	y;
